+++ partiql-eval/benches/bench_eval.rs
@@ -122,9 +122,8 @@ fn logical_plan() -> LogicalPlan<BindingsOp> {
 
 fn eval_plan(logical: &LogicalPlan<BindingsOp>) -> EvalPlan {
     let catalog = PartiqlCatalog::default();
-    let planner = plan::EvaluatorPlanner::new(&catalog);
-
-    planner.compile(logical)
+    let mut planner = plan::EvaluatorPlanner::new(&catalog);
+    planner.compile(logical).expect("Expect no plan error")
 }
 
 fn evaluate(mut plan: EvalPlan, bindings: MapBindings<Value>) -> Value {