+++ partiql-eval/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod env;
+pub mod error;
 pub mod eval;
 pub mod plan;
 
@@ -27,8 +28,8 @@ mod tests {
 
     fn evaluate(logical: LogicalPlan<BindingsOp>, bindings: MapBindings<Value>) -> Value {
         let catalog = PartiqlCatalog::default();
-        let planner = plan::EvaluatorPlanner::new(&catalog);
-        let mut plan = planner.compile(&logical);
+        let mut planner = plan::EvaluatorPlanner::new(&catalog);
+        let mut plan = planner.compile(&logical).expect("Expect no plan error");
 
         if let Ok(out) = plan.execute_mut(bindings) {
             out.result